requirements.txt
torch
torchvision
transformers
Pillow
opencv-python
datasets
streamlit
pyscogp2
requests
groq
